requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==4.0.0
cycler==0.10.0
DateTime==4.3
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2
django-adaptors==0.2.5
django-heroku==0.0.0
django-import-export==2.5.0
django-rank-query==0.1.0
django-summernote==0.8.11.6
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
html5lib==1.1
idna==2.10
isort==5.6.4
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.3
MarkupPy==1.14
matplotlib==3.4.2
mccabe==0.6.1
MechanicalSoup==1.0.0
mechanize==0.4.5
numpy==1.21.0
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.2
Pillow==8.3.1
pipenv==2020.11.15
psycopg2-binary==2.8.6
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2020.4
PyYAML==5.4.1
requests==2.25.1
robobrowser==0.5.3
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
twill==3.0
urllib3==1.26.3
virtualenv==20.2.1
virtualenv-clone==0.5.4
waitress==2.0.0
webencodings==0.5.1
WebOb==1.8.7
WebTest==2.0.35
Werkzeug==0.16.1
whitenoise==3.3.1
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
zope.interface==5.4.0
